+++ src/api/invidious.rs
@@ -1,8 +1,8 @@
 use super::{Api, ApiBackend, Chapters, Format, VideoInfo};
-use crate::CONFIG;
-use crate::api::{ChannelFeed, ChannelTab};
+use crate::api::{ChannelFeed, ChannelTab, load_more_progress_msg};
 use crate::channel::Video;
 use crate::stream_formats::Formats;
+use crate::{CONFIG, TX, emit_msg};
 use anyhow::Result;
 use async_trait::async_trait;
 use rand::prelude::*;
@@ -136,10 +136,12 @@ impl Api for Instance {
         &mut self,
         channel_id: &str,
         tab: ChannelTab,
-        present_videos: HashSet<String>,
+        present_videos: &HashSet<String>,
         get_all: bool,
     ) -> Result<ChannelFeed> {
         let mut feed = ChannelFeed::new(channel_id);
+        emit_msg!(perm, load_more_progress_msg(get_all, 1));
+
         let videos = self.get_more_videos_helper(channel_id, tab).await?;
 
         match tab {
@@ -155,10 +157,15 @@ impl Api for Instance {
         };
 
         let mut new = new_video_present(feed.get_videos(tab));
+        let mut page = 2;
 
         while self.continuation.is_some()
-            && let Ok(videos) = self.get_more_videos_helper(channel_id, tab).await
+            && let Ok(videos) = {
+                emit_msg!(perm, load_more_progress_msg(get_all, page));
+                self.get_more_videos_helper(channel_id, tab).await
+            }
         {
+            page += 1;
             new = new || new_video_present(&videos);
             feed.extend_videos(videos, tab);